chatapi/requirements.txt
argon2-cffi==20.1.0
asgi-redis==1.4.3
asgiref==1.1.2
cffi==1.14.4
Django==3.1.4
django-cors-headers==3.6.0
djangorestframework==3.12.2
msgpack-python==0.5.6
Pillow==8.0.1
pycparser==2.20
pytz==2020.4
redis==2.10.6
six==1.15.0
sqlparse==0.4.1
